Improve UI of login form.
This commit uses HTML formatted strings to show
the user which fields in the login form are required
to be able to login. However, the formatting doesn't
work for TextInputLayout, only for TextView. Therefore,
we toggle between setting the hint on the TextView (to
get color coding) or on the TextInputLayout (to get
floating hints).

This commit also adds some styling to the login button:
Specifically, it turns it pale, not just the background
but the text color as well until all required fields are entered.