'AC-571 - Fixing Application Crash'
## Description of what I changed
Added a condition to check whether the concepts in DB are 0 and disabled the user from clicking the Download Button if this condition is true and a toast will be displayed on page load informing the user of this. Further, null checks were introduced to ConceptDownloadService class.

## Issue I worked on
This is a critical bug in the Android Application
JIRA Issue: https://issues.openmrs.org/browse/AC-571

## Checklist: I completed these to help reviewers :)
- [x] My pull request only contains **ONE single commit**

- [x] I have **added tests** to cover my changes. (Only existing code was refactored)

- [x] All new and existing **tests passed**.

- [x] My pull request is **based on the latest changes** of the master branch.