Ignore to check if country codepicker is not empty
Better not to include line to check if the country code is not empty, As it always returning True value which make the whole method 'areFieldsNotEmpty()' to return true, By which it is always showing Dialogue box saying "All patient information entered will be lost", even if all the fields are empty. This might be due to default value of county box picker is selected.